Add filter that wraps the KdTreeBasedKmeansEstimator from ITK
This filter enables the use of the KdTreeBasedKMeansEstimator in ITK, which is
a more generic implementation of the other ITK k means filter.  The k-d tree
k means is faster but more memory inefficient.  This filter can operate on any
data type (not just the set default pixel type), but is currently only able
to classify data that are 3-vectors.